+++ RELEASENOTES.md
@@ -117,6 +117,7 @@ The latter links can be used for OTA upgrades too like ``OtaUrl https://ota.tasm
 - Support for JSON value pair `"ARCH"` in template being either ESP8266, ESP32, ESP32C2, ESP32C3, ESP32C5, ESP32C6, ESP32H2, ESP32H4, ESP32P4, ESP32S2 or ESP32S3
 - Support for Sensirion SCD42 and SCD43 CO2 sensor
 - Support for Sensirion STCC4 CO2 sensor
+- ESP8266 redesigned I2C Wire driver to support second I2C bus
 - Environment sensors CCS811, SGP30 and SGP40 second I2C bus support
 - Real Time Clocks BM8563, PCF85063 and PCF85363 second I2C bus support
 - I2S full duplex, auto rx sample rate [#24469](https://github.com/arendst/Tasmota/issues/24469)
@@ -129,7 +130,6 @@ The latter links can be used for OTA upgrades too like ``OtaUrl https://ota.tasm
 - Adafruit SGP40 library from v1.1.0 to v1.1.4
 - Sensirion Core library from v0.6.0 to v0.7.2
 - LVGL library from v9.4.0 to v9.5.0 [#24470](https://github.com/arendst/Tasmota/issues/24470)
-- ESP8266 redesigned I2C Wire driver to support second I2C bus
 - SHT1x software reset I2C bus after initial (un)detection
 - Matter improved parameters handling [#24471](https://github.com/arendst/Tasmota/issues/24471)